% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_para.R
\name{get_best_para}
\alias{get_best_para}
\title{Tune parameters}
\usage{
get_best_para(
  summs_exposure,
  summs_outcome,
  P_hat,
  snp.use,
  ncv = 3,
  ncores = 100,
  n_iter_cv = 30
)
}
\arguments{
\item{summs_exposure}{GWAS summary statistics for exposure, list, each element should be a data.frame including SNP, A1, A2, BETA, SE}

\item{summs_outcome}{GWAS summary statistics for outcome, a data.frame including SNP, A1, A2, BETA, SE}

\item{P_hat}{correlation matrix of exposures}

\item{snp.use}{rsid of the IVs}

\item{n_iter_cv}{Iterations in CV}

\item{n_cores}{Number of cores running in parallel}
}
\value{
best_parameters
}
\description{
A unified Bayesian framework to calculate PRS based on GWAS summary statistics
}
