% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{P_hat}
\alias{P_hat}
\title{Datasets
Demo GWAS data of 1 outcome and 6 expousres}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 6 columns.
}
\usage{
data(dat)
}
\description{
Datasets
Demo GWAS data of 1 outcome and 6 expousres
}
\keyword{datasets}
