% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mvmrpress.R
\name{run_mvmrpress}
\alias{run_mvmrpress}
\title{Main function}
\usage{
run_mvmrpress(
  summs_exposure,
  summs_outcome,
  snp.use,
  P_hat,
  para_theta = 0.01,
  para_gamma = 10,
  n_boots = 100,
  n_cores = 100,
  n_iter = 200,
  n_iter_boot = 100
)
}
\arguments{
\item{summs_exposure}{GWAS summary statistics for exposure, list, each element should be a data.frame including SNP, A1, A2, BETA, SE}

\item{summs_outcome}{GWAS summary statistics for outcome, a data.frame including SNP, A1, A2, BETA, SE}

\item{snp.use}{rsid of the IVs}

\item{P_hat}{correlation matrix of exposures}

\item{n_boots}{Number of boostraps}

\item{n_cores}{Number of cores running in parallel}

\item{n_iter}{Iterations}

\item{n_iter_boot}{Iterations in bootstrap}

\item{lambda_theta}{Parameter}

\item{mu_gamma}{Parameter}
}
\value{
result
}
\description{
MVMRPRESS
}
